Tried out bolthold/boltdb but it's significantly slower to save (weird) Added FleetWaypointTaskTile with server side update Changed primary keys to uint64 Moved discoverer into its own interface Changed all slices of objects to slices to pointers to objects to clean up the loops

+++ db/players_test.go
@@ -23,8 +23,8 @@ func TestDB_FindPlayerByGameId(t *testing.T) {
 
 	tests := []struct {
 		name    string
-		gameId  uint
-		userId  uint
+		gameId  uint64
+		userId  uint64
 		want    *game.Player
 		wantErr bool
 	}{
